requirements.txt
aiohttp>=3.9.0
asyncio
dnspython>=2.4.0
requests>=2.31.0
python-nmap>=0.7.1
shodan>=1.30.0
pyyaml>=6.0
reportlab>=4.0.0
jinja2>=3.1.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-whois>=0.8.0
cryptography>=41.0.0
pyjwt>=2.8.0
click>=8.1.0
rich>=13.7.0
tabulate>=0.9.0
colorama>=0.4.6

# AI/ML (optional - for future enhancement)
# torch>=2.1.0
# transformers>=4.35.0
# onnxruntime>=1.16.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.10.0
flake8>=6.1.0
